Update fingerprint.py
This is the code from an example.  In order this code to run we need multiple programs installed to the pi:
CircuitPython (Allows for us to control the scanner)
Adafruit CircuitPython Fingerprint (module used to run fingerprint scanner)
Adafruit_Blinka (This provides Curcuit python support in python)
Adafruit suggest using Mu to code, it is an IDE created specifically for this type of work and looked the simplest to work in

I have made a few minor changes that I hope will allow this code to run on the pi but I am not sure.